[Scrapers] Move movie scrapers into sub-directories for each provider
