+++ src/globals/ScraperManager.cpp
@@ -3,15 +3,15 @@
 #include "scrapers/ScraperInterface.h"
 #include "scrapers/concert/ConcertScraper.h"
 #include "scrapers/concert/tmdb/TmdbConcert.h"
-#include "scrapers/movie/AEBN.h"
-#include "scrapers/movie/AdultDvdEmpire.h"
-#include "scrapers/movie/CustomMovieScraper.h"
-#include "scrapers/movie/HotMovies.h"
-#include "scrapers/movie/IMDB.h"
 #include "scrapers/movie/MovieScraper.h"
-#include "scrapers/movie/OFDb.h"
-#include "scrapers/movie/TMDb.h"
-#include "scrapers/movie/VideoBuster.h"
+#include "scrapers/movie/adultdvdempire/AdultDvdEmpire.h"
+#include "scrapers/movie/aebn/AEBN.h"
+#include "scrapers/movie/custom/CustomMovieScraper.h"
+#include "scrapers/movie/hotmovies/HotMovies.h"
+#include "scrapers/movie/imdb/IMDB.h"
+#include "scrapers/movie/ofdb/OFDb.h"
+#include "scrapers/movie/tmdb/TMDb.h"
+#include "scrapers/movie/videobuster/VideoBuster.h"
 #include "scrapers/music/MusicScraper.h"
 #include "scrapers/music/UniversalMusicScraper.h"
 #include "scrapers/tv_show/TvScraper.h"
