+++ src/scrapers/image/FanartTvMusicArtists.cpp
@@ -10,7 +10,7 @@
 #include "data/Storage.h"
 #include "network/NetworkRequest.h"
 #include "scrapers/image/FanartTv.h"
-#include "scrapers/movie/TMDb.h"
+#include "scrapers/movie/tmdb/TMDb.h"
 
 namespace mediaelch {
 namespace scraper {
