+++ src/scrapers/image/TMDbImages.cpp
@@ -1,6 +1,6 @@
 #include "TMDbImages.h"
 
-#include "scrapers/movie/TMDb.h"
+#include "scrapers/movie/tmdb/TMDb.h"
 #include "settings/Settings.h"
 
 namespace mediaelch {
