+++ src/scrapers/movie/CMakeLists.txt
@@ -1,0 +1,8 @@
+add_subdirectory(adultdvdempire)
+add_subdirectory(aebn)
+add_subdirectory(hotmovies)
+add_subdirectory(imdb)
+add_subdirectory(ofdb)
+add_subdirectory(tmdb)
+add_subdirectory(videobuster)
+add_subdirectory(custom)
