+++ src/scrapers/movie/adultdvdempire/AdultDvdEmpire.cpp
