+++ src/scrapers/movie/hotmovies/HotMovies.cpp
