+++ src/scrapers/movie/ofdb/CMakeLists.txt
@@ -1,0 +1,7 @@
+add_library(mediaelch_scraper_movie_ofdb OBJECT OFDb.cpp)
+
+target_link_libraries(
+  mediaelch_scraper_movie_ofdb PRIVATE Qt5::Sql Qt5::Widgets Qt5::Multimedia
+                                       Qt5::Xml
+)
+mediaelch_post_target_defaults(mediaelch_scraper_movie_ofdb)
