+++ src/ui/movies/MovieMultiScrapeDialog.cpp
@@ -2,9 +2,9 @@
 #include "ui_MovieMultiScrapeDialog.h"
 
 #include "globals/Manager.h"
-#include "scrapers/movie/CustomMovieScraper.h"
-#include "scrapers/movie/IMDB.h"
-#include "scrapers/movie/TMDb.h"
+#include "scrapers/movie/custom/CustomMovieScraper.h"
+#include "scrapers/movie/imdb/IMDB.h"
+#include "scrapers/movie/tmdb/TMDb.h"
 #include "settings/Settings.h"
 #include "ui/small_widgets/MyCheckBox.h"
