+++ src/ui/movies/MovieSearchWidget.cpp
@@ -4,10 +4,10 @@
 #include "data/Locale.h"
 #include "globals/Manager.h"
 #include "globals/Meta.h"
-#include "scrapers/movie/CustomMovieScraper.h"
-#include "scrapers/movie/IMDB.h"
 #include "scrapers/movie/MovieScraper.h"
-#include "scrapers/movie/TMDb.h"
+#include "scrapers/movie/custom/CustomMovieScraper.h"
+#include "scrapers/movie/imdb/IMDB.h"
+#include "scrapers/movie/tmdb/TMDb.h"
 #include "settings/Settings.h"
 
 #include <QDebug>
