+++ src/ui/settings/ScraperSettingsWidget.cpp
@@ -3,9 +3,9 @@
 
 #include "globals/Manager.h"
 #include "scrapers/concert/ConcertScraper.h"
-#include "scrapers/movie/CustomMovieScraper.h"
-#include "scrapers/movie/IMDB.h"
 #include "scrapers/movie/MovieScraper.h"
+#include "scrapers/movie/custom/CustomMovieScraper.h"
+#include "scrapers/movie/imdb/IMDB.h"
 #include "scrapers/music/MusicScraper.h"
 #include "scrapers/tv_show/TvScraper.h"
 #include "scrapers/tv_show/thetvdb/TheTvDb.h"
