+++ test/scrapers/tvmaze/testTvMazeEpisodeScrapeJob.cpp
@@ -1,6 +1,6 @@
 #include "test/test_helpers.h"
 
-#include "scrapers/movie/IMDB.h"
+#include "scrapers/movie/imdb/IMDB.h"
 #include "scrapers/tv_show/tvmaze/TvMaze.h"
 #include "scrapers/tv_show/tvmaze/TvMazeEpisodeScrapeJob.h"
 #include "test/scrapers/tvmaze/testTvMazeHelper.h"
